+++ person.rb
@@ -1,4 +1,6 @@
-require_relative('./nameable')
+require './nameable'
+require './capitalize_decorator'
+require './trimmer_decorator'
 
 class Person < Nameable
   def initialize(age, name = 'Unknown', parent_permission: true)
